[Consensus] Dedicated consensus message verifier to reduce verification latency
Fix runtime to be async

name fix

Fix runtime
