+++ consensus/Cargo.toml
@@ -67,6 +67,7 @@ mirai-annotations = { workspace = true }
 move-core-types = { workspace = true }
 num-derive = { workspace = true }
 num-traits = { workspace = true }
+num_cpus = "1.16.0"
 once_cell = { workspace = true }
 rand = { workspace = true }
 rayon = { workspace = true }
