+++ consensus/src/consensus_provider.rs
@@ -39,8 +39,9 @@ pub fn start_consensus(
     aptos_db: DbReaderWriter,
     reconfig_events: ReconfigNotificationListener<DbBackedOnChainConfig>,
     vtxn_pool: VTxnPoolState,
-) -> (Runtime, Arc<StorageWriteProxy>, Arc<QuorumStoreDB>) {
+) -> (Runtime, Runtime, Arc<StorageWriteProxy>, Arc<QuorumStoreDB>) {
     let runtime = aptos_runtimes::spawn_named_runtime("consensus".into(), None);
+    let signature_verification_rt = aptos_runtimes::spawn_named_runtime("sig-verify".into(), None);
     let storage = Arc::new(StorageWriteProxy::new(node_config, aptos_db.reader.clone()));
     let quorum_store_db = Arc::new(QuorumStoreDB::new(node_config.storage.dir()));
 
@@ -77,6 +78,7 @@ pub fn start_consensus(
         quorum_store_db.clone(),
         reconfig_events,
         bounded_executor,
+        signature_verification_rt.handle().clone(),
         aptos_time_service::TimeService::real(),
         vtxn_pool,
     );
@@ -87,5 +89,5 @@ pub fn start_consensus(
     runtime.spawn(epoch_mgr.start(timeout_receiver, network_receiver));
 
     debug!("Consensus started.");
-    (runtime, storage, quorum_store_db)
+    (runtime, signature_verification_rt, storage, quorum_store_db)
 }
