+++ consensus/src/lib.rs
@@ -53,6 +53,7 @@ pub mod consensus_provider;
 /// Required by the telemetry service
 pub mod counters;
 mod execution_pipeline;
+mod msg_verifier;
 /// AptosNet interface.
 pub mod network_interface;
 mod payload_manager;
